+++ src/features/home/index.ts
@@ -2,4 +2,6 @@ export * from "./hooks/useHomeDashboard";
 export * from "./hooks/useHomeDashboardLayout";
 export * from "./pages/HomePage";
 export * from "./dashboardLayout";
+export * from "./personalInsights";
+export * from "./components/HomePersonalInsightsCard";
 export * from "./types";